requirements.txt
pathway==0.27.0
sentence-transformers==5.1.2
litellm==1.44.28
google-generativeai==0.8.5
python-dotenv==1.2.1
pandas==2.3.3
pyarrow==18.1.0
streamlit==1.51.0
scikit-learn==1.7.2
numpy==1.26.4
streamlit-autorefresh
